[sortinghat] Remove the unnecessary file
We are using poetry for managing the project and release-tools
for creating releases and publishing packages. GitHub Actions
are also configured for the CI replacing Travis CI.

This commit removes the unnecessary files from the repository.
It also adds the .flake8 configuration file which replaces the
functionality of setup.cfg file.

Signed-off-by: Venu Vardhan Reddy Tekula <venu@bitergia.com>