[log] Prevent to overwrite log objects when ids are reused
With the current code, transactions or operations objects
will be overwritten in the database if the same id is used.

This patch fixes the problem using the option 'force_insert'
on 'save()' calls. When this option is activated, it will
run 'INSERT' queries instead of 'UPDATE'. If an object already
has the same id, the engine will raise an IntegrityError
exception because the entity will be duplicated.

This patch adds the proper unit tests.

Signed-off-by: Santiago Dueñas <sduenas@bitergia.com>